stock/requirements.txt
kafka-python~=2.0.2
pytz~=2021.1
APScheduler~=3.7.0